Implement the parallel callback
This function is called as a method.  The '$data' passed as its first
argument is a deep cloned version of the invocant from the original
thread.  It must be cloned because Perl has no idea of 'const'.
However, this only happens once per worker and the worker may update
state in the invocant.  This is slightly different to the semantics of
parallel_for, which is supposed to have a 'const' body function.

The various variables that were explicit before are unpacked at the
start of the function, and the original inner loop function is called
over the range.

The concurrent containers are "lazy clone", but they don't cache the
cloned value, and will only update the "official" version on
re-assignment.  So it requires a fetch, working with the returned
value and re-assignment.  Locking slots is also possible, but not
necessary here.