Pick a closely matching body function API, and conform to it.
This script uses three main variables to communicate state between the
inner loop chosen for optimisation and the rest of the script; and
already passes around a data structure which is almost like an object.

So the "for_int_method" body function could work, if they were all put
inside of one of them, and then "blessed" to make it an object.