Create the body object
This is a method of the library initialization object; when the body
function runs it will load the modules that initializer wants before
starting.

The arguments are used to build the body object; in this case, the
object itself, and the name of the method to call to process a chunk.